software/pyproject.toml
[tool.pdm.version]
source = "scm"

[project]
dynamic = ["version"]

name = "glasgow"
description = "Software for the Glasgow Interface Explorer, a digital interface multitool"
authors = [
  {name = "Catherine", email = "whitequark@whitequark.org"},
  {name = "Glasgow Interface Explorer contributors"}
]
license = {text = "0BSD OR Apache-2.0"}
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved",
  "Topic :: Software Development :: Embedded Systems",
  "Topic :: System :: Hardware",
]

# See the source of `support.plugin` for reasoning behind this constraint.
requires-python = "~=3.9, !=3.10.0, !=3.10.1, !=3.10.2"

# Glasgow has only a few carefully chosen dependencies. The rationale for having them, as well as
# the rationale for the particular version constraint, is explained below, and should be kept
# synchronized with the source code. For clarity, we do not use the `~=` version constraint.
#
# Whenever dependencies are changed in any way, run `pdm run update-pdm.min.lock`. If you do not
# run this command, the CI will reject the pull request with the change.
dependencies = [
  # We use `typing` features not available in the lowest Python version we support. The library
  # `typing_extensions` provides shims for such features. It uses SemVer.
  "typing_extensions>=4.8,<5",
  # Amaranth is the core of the Glasgow software/gateware interoperability layer. It uses SemVer.
  # We use the git version of Amaranth until the next release.
  "amaranth @ git+https://github.com/amaranth-lang/amaranth.git@4e4085a95b",
  # `packaging` is used in the plugin system, `support.plugin`. It uses CalVer: the major version
  # is the two last digits of the year and the minor version is the release within that year.
  "packaging>=23.0",
  # `platformdirs` is used in the bitstream builder to use platform-appropriate cache directories.
  # It uses SemVer.
  "platformdirs>=3.0.0,<4",
  # `fx2` is effectively maintained together with Glasgow. It uses SemVer, and keeps backward
  # compatibility within the 0.x release series.
  "fx2>=0.11,<1",
  # `libusb1` is used to communicate with the device, and its API mirrors the stable API/ABI of
  # of the native `libusb1` library. It increases major version when dropping support for older
  # Python versions.
  "libusb1>=3.1.0",
  # `pyvcd` is used in the applet analyzer to dump waveform files. It is also a dependency of
  # Amaranth, and the version range here must be compatible with Amaranth's.
  "pyvcd>=0.2,<0.5",
]

[project.optional-dependencies]
# By default, Glasgow uses the YoWASP (https://yowasp.org/) to build bitstreams. YoWASP is not
# installable on certain architectures and platforms, so this dependency is optional.
builtin-toolchain = [
  # The version of `amaranth-yosys` is constrained by Amaranth; this optional dependency only
  # includes it in the dependency resolution set.
  "amaranth-yosys",
  # `yowasp-runtime` is a dependency of other toolchain components, and it is constrained here to
  # the lowest that has features transitively required by Glasgow.
  "yowasp-runtime>=1.42",
  # Most versions of Yosys and nextpnr work; the minimum versions listed here are known good.
  "yowasp-yosys>=0.31.0.13",
  "yowasp-nextpnr-ice40>=0.1",
]

# The rest of the optional dependences belong to plugins (applets, loggers, etc.)
http = [
  "aiohttp>=3.9.0b0",
  "aiohttp-remotes>=1.2,<2",
]

[project.scripts]
glasgow = "glasgow.cli:main"

[project.entry-points."glasgow.applet"]
selftest = "glasgow.applet.internal.selftest:SelfTestApplet"
benchmark = "glasgow.applet.internal.benchmark:BenchmarkApplet"

analyzer = "glasgow.applet.interface.analyzer:AnalyzerApplet"
better-la = "glasgow.applet.interface.better_la:BetterLAApplet"
uart = "glasgow.applet.interface.uart:UARTApplet"
spi-controller = "glasgow.applet.interface.spi_controller:SPIControllerApplet"
i2c-initiator = "glasgow.applet.interface.i2c_initiator:I2CInitiatorApplet"
i2c-target = "glasgow.applet.interface.i2c_target:I2CTargetApplet"
jtag-pinout = "glasgow.applet.interface.jtag_pinout:JTAGPinoutApplet"
jtag-probe = "glasgow.applet.interface.jtag_probe:JTAGProbeApplet"
jtag-openocd = "glasgow.applet.interface.jtag_openocd:JTAGOpenOCDApplet"
jtag-svf = "glasgow.applet.interface.jtag_svf:JTAGSVFApplet"
ps2-host = "glasgow.applet.interface.ps2_host:PS2HostApplet"
sbw-probe = "glasgow.applet.interface.sbw_probe:SpyBiWireProbeApplet"

memory-24x = "glasgow.applet.memory._24x:Memory24xApplet"
memory-25x = "glasgow.applet.memory._25x:Memory25xApplet"
memory-onfi = "glasgow.applet.memory.onfi:MemoryONFIApplet"
memory-prom = "glasgow.applet.memory.prom:MemoryPROMApplet"
memory-floppy = "glasgow.applet.memory.floppy:MemoryFloppyApplet"

debug-arc = "glasgow.applet.debug.arc:DebugARCApplet"
debug-arm = "glasgow.applet.debug.arm.jtag:DebugARMJTAGApplet"
debug-mips = "glasgow.applet.debug.mips:DebugMIPSApplet"

program-avr-spi = "glasgow.applet.program.avr.spi:ProgramAVRSPIApplet"
program-ice40-flash = "glasgow.applet.program.ice40_flash:ProgramICE40FlashApplet"
program-ice40-sram = "glasgow.applet.program.ice40_sram:ProgramICE40SRAMApplet"
program-m16c = "glasgow.applet.program.m16c:ProgramM16CApplet"
program-mec16xx = "glasgow.applet.program.mec16xx:ProgramMEC16xxApplet"
program-nrf24lx1 = "glasgow.applet.program.nrf24lx1:ProgramNRF24Lx1Applet"
program-stusb4500-nvm = "glasgow.applet.program.stusb4500_nvm:StUsb4500NvmApplet"
program-xc6s = "glasgow.applet.program.xc6s:ProgramXC6SApplet"
program-xc9500xl = "glasgow.applet.program.xc9500xl:ProgramXC9500XLApplet"

control-tps6598x = "glasgow.applet.control.tps6598x:ControlTPS6598xApplet"

sensor-bmx280 = "glasgow.applet.sensor.bmx280:SensorBMx280Applet"
sensor-hx711 = "glasgow.applet.sensor.hx711:SensorHX711Applet"
sensor-ina260 = "glasgow.applet.sensor.ina260:SensorINA260Applet"
sensor-mouse-ps2 = "glasgow.applet.sensor.mouse_ps2:SensorMousePS2Applet"
sensor-pmsx003 = "glasgow.applet.sensor.pmsx003:SensorPMSx003Applet"
sensor-sen5x = "glasgow.applet.sensor.sen5x:SensorSEN5xApplet"
sensor-scd30 = "glasgow.applet.sensor.scd30:SensorSCD30Applet"

display-hd44780 = "glasgow.applet.display.hd44780:DisplayHD44780Applet"
display-pdi = "glasgow.applet.display.pdi:DisplayPDIApplet"

audio-dac = "glasgow.applet.audio.dac:AudioDACApplet"
audio-yamaha-opx = "glasgow.applet.audio.yamaha_opx:AudioYamahaOPxApplet [http]"

video-hub75-output = "glasgow.applet.video.hub75_output:VideoHub75OutputApplet"
video-rgb-input = "glasgow.applet.video.rgb_input:VideoRGBInputApplet"
video-vga-output = "glasgow.applet.video.vga_output:VGAOutputApplet"
video-ws2812-output = "glasgow.applet.video.ws2812_output:VideoWS2812OutputApplet"

radio-nrf24l01 = "glasgow.applet.radio.nrf24l01:RadioNRF24L01Applet"

[project.urls]
"Documentation" = "https://glasgow-embedded.org/"
"Source Code" = "https://github.com/GlasgowEmebedded/Glasgow"
"Bug Tracker" = "https://github.com/GlasgowEmebedded/Glasgow/issues"

# Build system configuration

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

# Development workflow configuration

[tool.pdm.scripts]
"update-pdm.min.lock".cmd = "pdm lock -L pdm.min.lock -S direct_minimal_versions"
test.cmd = "python -m unittest"
